Move MongoModel trait implementation into db files
