+++ homelab_backend/src/models/log/log_db.rs
@@ -1,10 +1,24 @@
-use mongodb::bson::{doc, Bson};
-
 use crate::{
-    db::{str_to_object_id, PatDatabase},
+    db::{str_to_object_id, MongoModel, PatDatabase},
     error_handler::DbError,
     models::log::Log,
 };
+use mongodb::bson::{doc, oid::ObjectId, Bson};
+
+impl MongoModel for Log {
+    fn collection_name() -> &'static str {
+        "logs"
+    }
+    fn model_name() -> &'static str {
+        "Log"
+    }
+    fn mongo_id(&self) -> Result<ObjectId, DbError> {
+        match self.id.parse::<ObjectId>() {
+            Ok(res) => Ok(res),
+            Err(_) => Err(DbError::BadId),
+        }
+    }
+}
 
 pub async fn db_get_logs_for_user(db_handle: &PatDatabase, user_id: String) -> Result<Vec<Log>, DbError> {
     let doc = doc! { "user_id": user_id };
