+++ homelab_backend/src/models/log/mod.rs
@@ -1,9 +1,7 @@
 pub mod log_db;
 
 use super::deserialize_id;
-use crate::{db::MongoModel, error_handler::DbError};
 use hyper::body::Bytes;
-use mongodb::bson::oid::ObjectId;
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug)]
@@ -14,21 +12,6 @@ pub struct Log {
     pub uri: String,
     pub user_id: String,
     date_time: i64,
-}
-
-impl MongoModel for Log {
-    fn collection_name() -> &'static str {
-        "logs"
-    }
-    fn model_name() -> &'static str {
-        "Log"
-    }
-    fn mongo_id(&self) -> Result<ObjectId, DbError> {
-        match self.id.parse::<ObjectId>() {
-            Ok(res) => Ok(res),
-            Err(_) => Err(DbError::BadId),
-        }
-    }
 }
 
 // TODO: Should have a log retention policy/task which auto deletes old tasks.
